+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ### Improvements
 
 - Make the encoding priority view undoable.
+- [trivial] Enable delete selected separators in the encoding priority view with the Delete key.
 
 
 ### Fixes